update geo_modifs* entities id type #14313

+++ src/main/java/fr/microtec/geo2/persistance/entity/common/GeoModification.java
@@ -1,5 +1,6 @@
 package fr.microtec.geo2.persistance.entity.common;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -34,7 +35,7 @@ public class GeoModification {
 					@org.hibernate.annotations.Parameter(name = "isSequence", value = "true")
 			}
 	)
-	private Integer id;
+	private BigDecimal id;
 
 	@OneToMany(fetch = FetchType.LAZY, mappedBy = "modification")
 	private List<GeoModificationCorps> corps;
